Fix: Fix CalendarViewModel
